Re organize the package
